Start implementation of using a volume for model cache
From #47, this is a rough draft of a possible API for specifying the
models needed to be pulled, and the storage resource needed to store
them.

This does not currently implement the volume portion. It also does
nothing with the models specified.

What it does, however, is to add an emptyDir volume to the container,
mounted at /opt/docling/models.

Signed-off-by: Brad P. Crochet <brad@redhat.com>